Removed the following modules which are not compatible with 2.2:
- blocks/facetoface
- mod/book
- mod/certificate
- mod/facetoface
- mod/hotpot
- mod/lams

+++ admin/timezone.php
@@ -1,8 +1,13 @@
-<?php   // $Id$
+<?php
 
     require_once('../config.php');
 
-    $zone = optional_param('zone', '', PARAM_PATH); //not a path, but it looks like it anyway
+    $zone = optional_param('zone', '', PARAM_RAW);
+
+    if (!is_numeric($zone)) {
+         //not a path, but it looks like it anyway
+         $zone = clean_param($zone, PARAM_PATH);
+    }
 
     require_login();
 
@@ -13,30 +18,36 @@
     $strusers = get_string("users");
     $strall = get_string("all");
 
-    print_header($strtimezone, $strtimezone, build_navigation(array(array('name' => $strtimezone, 'link' => null, 'type' => 'misc'))));
+    $PAGE->set_url('/admin/timezone.php');
+    $PAGE->set_title($strtimezone);
+    $PAGE->set_heading($strtimezone);
+    $PAGE->navbar->add($strtimezone);
+    echo $OUTPUT->header();
 
-    print_heading("");
+    echo $OUTPUT->heading("");
 
-    if (!empty($zone) and confirm_sesskey()) {
-        $db->debug = true;
+    if (data_submitted() and !empty($zone) and confirm_sesskey()) {
         echo "<center>";
-        execute_sql("UPDATE {$CFG->prefix}user SET timezone = '$zone'");
-        $db->debug = false;
+        $DB->execute("UPDATE {user} SET timezone = ?", array($zone));
         echo "</center>";
 
         $USER->timezone = $zone;
+        $current = $zone;
+        echo $OUTPUT->notification('Timezone of all users changed', 'notifysuccess');
+    } else {
+        $current = 99;
     }
 
     require_once($CFG->dirroot.'/calendar/lib.php');
     $timezones = get_list_of_timezones();
 
-    echo '<center><form action="timezone.php" method="get">';
+    echo '<center><form action="timezone.php" method="post">';
     echo "$strusers ($strall): ";
-    choose_from_menu ($timezones, "zone", 99, get_string("serverlocaltime"), "", "99");
-    echo "<input type=\"hidden\" name=\"sesskey\" value=\"$USER->sesskey\" />";
-    echo "<input type=\"submit\" value=\"$strsavechanges\" />";
+    echo html_writer::select($timezones, "zone", $current, array('99'=>get_string("serverlocaltime")));
+    echo "<input type=\"hidden\" name=\"sesskey\" value=\"".sesskey()."\" />";
+    echo '<input type="submit" value="'.s($strsavechanges).'" />';
     echo "</form></center>";
 
-    print_footer();
+    echo $OUTPUT->footer();
 
-?>
+
